pyproject.toml
[project]
name = "mysql-php"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mysql-connector-python>=9.0",
    "pymysql>=1.1.2",
    "pymongo>=4.0",
]
